Des bricoles

+++ core/src/clock.rs
@@ -303,7 +303,7 @@ impl Clock {
     /// Start/stop synchronization feature in Ableton Link.
     pub fn play_pause(&mut self) {
         self.session_state
-            .set_is_playing(!self.is_playing(), self.micros());
+            .set_is_playing(!self.is_playing(), self.micros() as i64);
         self.commit_app_state();
     }
 
@@ -314,7 +314,7 @@ impl Clock {
 
     pub fn set_playing(&mut self, playing: bool) {
         self.session_state
-            .set_is_playing(playing, self.micros());
+            .set_is_playing(playing, self.micros() as i64);
         self.commit_app_state();
     }
